refactor out AddAtt to own class
